feat(board): unify Room model for DB and API layers
- Enable `utoipa` chrono feature for automatic OpenAPI schema generation
- Remove redundant `RoomResponse` struct and related conversion code
- Simplify time field serialization by leveraging chrono's default ISO 8601 format
- Update dependencies and documentation to reflect best practices
- Fix enum usage in tests to match defined types

+++ crates/board/tests/room_repository_tests.rs
@@ -27,7 +27,7 @@ fn create_test_room(name: &str) -> Room {
         id: None,
         name: name.to_string(),
         password: Some("test_password".to_string()),
-        status: RoomStatus::Open as i64,
+        status: RoomStatus::Open,
         max_size: 100,
         current_size: 0,
         max_times_entered: 1000,
